agentinstall_server/requirements.txt
mysql-connector-python
pyyaml
configparser